Fixed bug: jquery search highlighting
Now, when a user searches for "bass quality", both "bass"
and "quality" will be highlighted in the reviews.

Needed to use a different version of the jquery-highlighting plugin

A further step would be to allow the user to use quotations if they
DID want an exact match for "bass quality".